Move config API to subproject
