+++ build.gradle
@@ -61,7 +61,6 @@ dependencies {
 
 	implementation 'org.squiddev:cctweaks-lua:1.5.0-pr4'
 
-	implementation 'com.google.guava:guava:22.0'
 	implementation 'com.google.code.gson:gson:2.8.1'
 
 	testCompile 'junit:junit:4.12'
