+++ plugin-api/build.gradle
@@ -9,4 +9,6 @@ dependencies {
 	api 'dan200.computercraft:ComputerCraft:1.80pr1-build4'
 
 	api 'org.slf4j:slf4j-api:1.7.25'
+	
+	api 'com.google.guava:guava:22.0'
 }
