+++ plugin-api/src/main/java/net/clgd/ccemux/api/config/ConfigEntry.java
@@ -1,4 +1,4 @@
-package net.clgd.ccemux.config;
+package net.clgd.ccemux.api.config;
 
 import java.util.Optional;
