+++ plugin-api/src/main/java/net/clgd/ccemux/api/config/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * This package contains the api for the modular configuration system used by
+ * CCEmuX. Any type can be stored, although "basic" types (boxed primitives,
+ * strings, lists, etc) should be preferred for compatibility reasons.
+ * 
+ * @see Config
+ * @see ConfigProperty
+ * @see Group
+ */
+package net.clgd.ccemux.api.config;
