+++ src/main/java/net/clgd/ccemux/plugins/PluginManager.java
@@ -1,16 +1,28 @@
 package net.clgd.ccemux.plugins;
 
 import java.io.File;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.ServiceLoader;
 import java.util.function.Consumer;
 
 import lombok.val;
 import lombok.extern.slf4j.Slf4j;
-import net.clgd.ccemux.config.ConfigProperty;
-import net.clgd.ccemux.config.Group;
-import net.clgd.ccemux.emulation.*;
+import net.clgd.ccemux.api.config.ConfigProperty;
+import net.clgd.ccemux.api.config.Group;
+import net.clgd.ccemux.emulation.CCEmuX;
+import net.clgd.ccemux.emulation.EmuConfig;
+import net.clgd.ccemux.emulation.EmulatedComputer;
 import net.clgd.ccemux.emulation.filesystem.VirtualDirectory;
-import net.clgd.ccemux.plugins.hooks.*;
+import net.clgd.ccemux.plugins.hooks.Closing;
+import net.clgd.ccemux.plugins.hooks.ComputerCreated;
+import net.clgd.ccemux.plugins.hooks.ComputerRemoved;
+import net.clgd.ccemux.plugins.hooks.CreatingComputer;
+import net.clgd.ccemux.plugins.hooks.CreatingROM;
+import net.clgd.ccemux.plugins.hooks.Hook;
+import net.clgd.ccemux.plugins.hooks.InitializationCompleted;
+import net.clgd.ccemux.plugins.hooks.RendererCreated;
+import net.clgd.ccemux.plugins.hooks.Tick;
 import net.clgd.ccemux.rendering.Renderer;
 
 @Slf4j
