+++ src/main/java/net/clgd/ccemux/plugins/builtin/JFXPlugin.java
@@ -1,12 +1,14 @@
 package net.clgd.ccemux.plugins.builtin;
 
-import java.util.*;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Optional;
 
 import com.google.auto.service.AutoService;
 
 import net.clgd.ccemux.api.OperatingSystem;
-import net.clgd.ccemux.config.ConfigProperty;
-import net.clgd.ccemux.config.Group;
+import net.clgd.ccemux.api.config.ConfigProperty;
+import net.clgd.ccemux.api.config.Group;
 import net.clgd.ccemux.emulation.EmuConfig;
 import net.clgd.ccemux.plugins.Plugin;
 import net.clgd.ccemux.rendering.RendererFactory;
